+++ lib/VWOClient.ts
@@ -31,7 +31,7 @@ import { ContextModel } from './models/user/ContextModel';
 import HooksService from './services/HooksService';
 import { UrlUtil } from './utils/UrlUtil';
 
-import { getType, isBoolean, isNumber, isObject, isString } from './utils/DataTypeUtil';
+import { getType, isObject, isString, isBoolean, isNumber } from './utils/DataTypeUtil';
 
 import { buildMessage } from './utils/LogMessageUtil';
 import { Deferred } from './utils/PromiseUtil';
@@ -57,6 +57,7 @@ export interface IVWOClient {
     attributeValue: boolean | string | number,
     context: Record<string, any>,
   ): Promise<void>;
+  setAttribute(attributes: Record<string, boolean | string | number>, context: Record<string, any>): Promise<void>;
   updateSettings(settings?: Record<string, any>, isViaWebhook?: boolean): Promise<void>;
 }
 
@@ -252,73 +253,126 @@ export class VWOClient implements IVWOClient {
   }
 
   /**
-   * Sets an attribute for a user in the context provided.
+   * Sets an attribute or multiple attributes for a user in the provided context.
    * This method validates the types of the inputs before proceeding with the API call.
+   * There are two cases handled:
+   * 1. When attributes are passed as a map (key-value pairs).
+   * 2. When a single attribute (key-value) is passed.
    *
-   * @param {string} attributeKey - The key of the attribute to set.
-   * @param {string} attributeValue - The value of the attribute to set.
-   * @param {ContextModel} context - The context in which the attribute should be set, must include a valid user ID.
+   * @param {string | Record<string, boolean | string | number>} attributeOrAttributes - Either a single attribute key (string) and value (boolean | string | number),
+   *                                                                                        or a map of attributes with keys and values (boolean | string | number).
+   * @param {boolean | string | number | Record<string, any>} [attributeValueOrContext] - The value for the attribute in case of a single attribute, or the context when multiple attributes are passed.
+   * @param {Record<string, any>} [context] - The context which must include a valid user ID. This is required if multiple attributes are passed.
    */
   async setAttribute(
-    attributeKey: string,
-    attributeValue: boolean | string | number,
-    context: Record<string, any>,
+    attributeOrAttributes: string | Record<string, boolean | string | number>,
+    attributeValueOrContext?: boolean | string | number | Record<string, any>,
+    context?: Record<string, any>,
   ): Promise<void> {
     const apiName = 'setAttribute';
 
     try {
-      // Log the API call
-      LogManager.Instance.debug(
-        buildMessage(DebugLogMessagesEnum.API_CALLED, {
-          apiName,
-        }),
-      );
-
-      // Validate attributeKey is a string
-      if (!isString(attributeKey)) {
-        LogManager.Instance.error(
-          buildMessage(ErrorLogMessagesEnum.API_INVALID_PARAM, {
+      if (isObject(attributeOrAttributes)) {
+        // Log the API call
+        LogManager.Instance.debug(
+          buildMessage(DebugLogMessagesEnum.API_CALLED, {
             apiName,
-            key: 'attributeKey',
-            type: getType(attributeKey),
-            correctType: 'string',
           }),
         );
 
-        throw new TypeError('TypeError: attributeKey should be a string');
+        if (Object.entries(attributeOrAttributes).length < 1) {
+          LogManager.Instance.error(
+            buildMessage('Attributes map must contain atleast 1 key-value pair', {
+              apiName,
+              key: 'attributes',
+              type: getType(attributeOrAttributes),
+              correctType: 'object',
+            }),
+          );
+          throw new TypeError('TypeError: Attributes should be an object containing atleast 1 key-value pair');
+        }
+
+        // Case where multiple attributes are passed as a map
+        const attributes = attributeOrAttributes as Record<string, boolean | string | number>; // Type assertion
+
+        // Validate attributes is an object
+        if (!isObject(attributes)) {
+          throw new TypeError('TypeError: attributes should be an object containing key-value pairs');
+        }
+
+        // Validate that each attribute value is of a supported type
+        Object.entries(attributes).forEach(([key, value]) => {
+          if (typeof value !== 'boolean' && typeof value !== 'string' && typeof value !== 'number') {
+            LogManager.Instance.error(
+              buildMessage(ErrorLogMessagesEnum.API_INVALID_PARAM, {
+                apiName,
+                key,
+                type: getType(value),
+                correctType: ' boolean, string or number',
+              }),
+            );
+            throw new TypeError(
+              `Invalid attribute type for key "${key}". Expected boolean, string or number, but got ${getType(value)}`,
+            );
+          }
+
+          // Reject arrays and objects explicitly
+          if (Array.isArray(value) || (typeof value === 'object' && value !== null)) {
+            LogManager.Instance.error(
+              buildMessage(ErrorLogMessagesEnum.API_INVALID_PARAM, {
+                apiName,
+                key,
+                type: getType(value),
+                correctType: ' boolean | string | number | null',
+              }),
+            );
+            throw new TypeError(`Invalid attribute value for key "${key}". Arrays and objects are not supported.`);
+          }
+        });
+
+        // If we have only two arguments (attributeMap and context)
+        if (!context && attributeValueOrContext) {
+          context = attributeValueOrContext as Record<string, any>; // Assign context explicitly
+        }
+
+        // Validate user ID is present in context
+        if (!context || !context.id) {
+          LogManager.Instance.error(ErrorLogMessagesEnum.API_CONTEXT_INVALID);
+        }
+
+        const contextModel = new ContextModel().modelFromDictionary(context);
+        // Proceed with setting the attributes if validation is successful
+        await new SetAttributeApi().setAttribute(this.settings, attributes, contextModel);
+      } else {
+        // Case where a single attribute (key-value) is passed
+        const attributeKey = attributeOrAttributes;
+        const attributeValue = attributeValueOrContext;
+
+        // Validate attributeKey is a string
+        if (!isString(attributeKey)) {
+          throw new TypeError('attributeKey should be a string');
+        }
+
+        // Validate attributeValue is of valid type
+        if (!isBoolean(attributeValue) && !isString(attributeValue) && !isNumber(attributeValue)) {
+          throw new TypeError('attributeValue should be a boolean, string, or number');
+        }
+
+        // Validate user ID is present in context
+        if (!context || !context.id) {
+          throw new TypeError('Invalid context');
+        }
+
+        const contextModel = new ContextModel().modelFromDictionary(context);
+
+        // Create a map from the single attribute key-value pair
+        const attributeMap = { [attributeKey]: attributeValue };
+
+        // Proceed with setting the attribute map if validation is successful
+        await new SetAttributeApi().setAttribute(this.settings, attributeMap, contextModel);
       }
-      // Validate attributeValue is a string
-      if (!isString(attributeValue) && !isNumber(attributeValue) && !isBoolean(attributeValue)) {
-        LogManager.Instance.error(
-          buildMessage(ErrorLogMessagesEnum.API_INVALID_PARAM, {
-            apiName,
-            key: 'attributeValue',
-            type: getType(attributeValue),
-            correctType: 'boolean | string | number',
-          }),
-        );
-
-        throw new TypeError('TypeError: attributeValue should be a string');
-      }
-
-      // Validate user ID is present in context
-      if (!context || !context.id) {
-        LogManager.Instance.error(ErrorLogMessagesEnum.API_CONTEXT_INVALID);
-        throw new TypeError('TypeError: Invalid context');
-      }
-
-      const contextModel = new ContextModel().modelFromDictionary(context);
-
-      // Proceed with setting the attribute if validation is successful
-      await new SetAttributeApi().setAttribute(this.settings, attributeKey, attributeValue, contextModel);
     } catch (err) {
-      // Log any errors encountered during the operation
-      LogManager.Instance.info(
-        buildMessage(ErrorLogMessagesEnum.API_THROW_ERROR, {
-          apiName,
-          err,
-        }),
-      );
+      LogManager.Instance.info(buildMessage(ErrorLogMessagesEnum.API_THROW_ERROR, { apiName, err }));
     }
   }