+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2024-02-13
+
+### Added
+
+- Support for `Object` in `setAttribute` method to send multiple attributes at once.
+
+  ```javascript
+  const attributes = { attr1: value1, attr2: value2 };
+  client.setAttribute(attributes, context);
+  ```
+
 ## [1.11.0] - 2024-12-20
 
 ### Added